+++ src/hit.rs
@@ -8,7 +8,7 @@ pub struct HitRecord {
     pub p: Point3,
     pub normal: Vec3,
     pub mat: Rc<dyn Scatter>,
-    pub t: f64,
+    pub t: f32,
     pub front_face: bool,
 }
 
@@ -24,13 +24,13 @@ impl HitRecord {
 }
 
 pub trait Hit {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord>;
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord>;
 }
 
 pub type World = Vec<Box<dyn Hit>>;
 
 impl Hit for World {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let mut tmp_rec = None;
         let mut closest_so_far = t_max;
